requirement.txt
TensorFlow==1.12
sklearn
pyyaml
keras==2.2.5
argparse